Move nav components to separate files
