+++ homepage/design-system/src/app/components/organisms/navigation/index.ts
@@ -1,0 +1,2 @@
+export * from "./Navigation";
+export * from "./types";
